Fixed bugs and change balance
- fixed UI Responsive
- change balance of Villager, Palisade Gate, Lumber Camp, Mining Cart and Wind Mill

+++ Assets/Scripts/Object Database/CSharpScript/EconomicUpgradeDatabase.cs
@@ -13,7 +13,7 @@ public class EconomicUpgradeDatabase : ScriptableObject
 public class EconomicUpgradeInfo
 {
     public string upgradeType;
-    public int increaseGatheringSpeed;
+    public float increaseGatheringSpeed;
     public int increaseGatheringCapacity;
     public int goldCostForUpgradeGatheringSpeed;
     public int goldCostForUpgradeGatheringCapacity;
